update assessment utility calculation and reuse in various service component
