+++ src/app/shared/reports/calculations/utilityCalculation.ts
@@ -3,6 +3,7 @@ import { UtilityOptions } from "../../constants/utilityTypes";
 import { ConvertValue } from "../../conversions/convertValue";
 import { IdbAssessment } from "src/app/models/assessment";
 import { UtilityEnergyUse } from "src/app/models/utilityEnergyUses";
+import { UnitSettings } from "src/app/models/unitSettings";
 
 export function updateFacilityUtilityUseCost(facility: IdbFacility, companyEnergyUnit: string): IdbFacility {
   const convertValue = new ConvertValue();
@@ -48,11 +49,69 @@ export function updateFacilityUtilityUseCost(facility: IdbFacility, companyEnerg
   return facility;
 }
 
-export function updateAssessmentUtilityUseCostSavings(assessment: IdbAssessment, facility: IdbFacility, companyEnergyUnit: string): IdbAssessment {
-  const facilityUnitSettings = facility.unitSettings;
+export function updateAssessmentUtilityUseCostSavings(assessment: IdbAssessment, facilityUnitSettings: UnitSettings, companyEnergyUnit: string): IdbAssessment {
+  assessment = calculateAssessmentUtilityUseSavings(assessment, companyEnergyUnit);
+  assessment = calculateAssessmentUtilityCostSavings(assessment, facilityUnitSettings);
+  return assessment;
+}
+
+export function calculateAssessmentUtilityUseSavings(assessment: IdbAssessment, companyEnergyUnit: string): IdbAssessment {
   const convertValue = new ConvertValue();
-  let energyUse = 0, energyCost = 0, waterCost = 0;
+  let energyUse = 0;
   let energySavings = 0, waterSavings = 0;
+  assessment.utilityTypes.forEach(utilityType => {
+    let utilityEnergyUse: UtilityEnergyUse = assessment.utilityEnergyUses.find(
+      _energyUse => _energyUse.utilityType == utilityType);
+    if (utilityEnergyUse.include) {
+      let trimmedType = utilityType.replace(/\s+/g, ''); // Remove spaces
+      let camelCaseType = trimmedType.charAt(0).toLowerCase() + trimmedType.slice(1);
+      let convertedUse = 0, convertedUseForCost = 0;
+      let convertedSaving = 0, convertedSavingForCost = 0;
+      if (utilityType == 'Water' || utilityType == 'Waste Water') {
+        // calculate saving
+        convertedSavingForCost = convertValue.convertValue(
+          utilityEnergyUse.utilitySaving,
+          utilityEnergyUse.energyUnit,
+          'kgal').convertedValue; // default to kgal
+        if (isNaN(convertedSavingForCost)) {
+          convertedSavingForCost = 0;
+        }
+        waterSavings += convertedSavingForCost;
+      } else {
+        let selectedUtilityOption = UtilityOptions.find(
+          _option => _option.utilityType == utilityType);
+        let selectedUnitOption = selectedUtilityOption.energyUnitOptions.find(
+          _unitOption => _unitOption.value == utilityEnergyUse.energyUnit);
+        // calculate use
+        if (selectedUtilityOption.isStandardEnergyUnit
+          && selectedUnitOption?.isStandard !== false) {
+          convertedUse = convertValue.convertValue(
+            utilityEnergyUse.energyUse,
+            utilityEnergyUse.energyUnit,
+            companyEnergyUnit).convertedValue;
+        } else {
+          convertedUse = convertValue.convertValue(
+            utilityEnergyUse.energyUse * utilityEnergyUse.energyHHV,
+            utilityEnergyUse.energyUnitStandard,
+            companyEnergyUnit).convertedValue;
+        }
+        if (isNaN(convertedUse)) {
+          convertedUse = 0;
+        }
+        energyUse += convertedUse;
+        energySavings += convertedSaving;
+      }
+    }
+  });
+  assessment.energyUse = energyUse;
+  assessment.energySavings = energySavings;
+  assessment.waterSavings = waterSavings;
+  return assessment;
+}
+
+export function calculateAssessmentUtilityCostSavings(assessment: IdbAssessment, facilityUnitSettings: UnitSettings): IdbAssessment {
+  const convertValue = new ConvertValue();
+  let energyCost = 0, waterCost = 0;
   let energyCostSavings = 0, waterCostSavings = 0;
   assessment.utilityTypes.forEach(utilityType => {
     let utilityEnergyUse: UtilityEnergyUse = assessment.utilityEnergyUses.find(
@@ -80,30 +139,12 @@ export function updateAssessmentUtilityUseCostSavings(assessment: IdbAssessment,
         if (isNaN(convertedSavingForCost)) {
           convertedSavingForCost = 0;
         }
-        waterSavings += convertedSavingForCost;
         waterCostSavings += convertedSavingForCost * facilityUnitSettings[`${camelCaseType}Price`];
       } else {
         let selectedUtilityOption = UtilityOptions.find(
           _option => _option.utilityType == utilityType);
         let selectedUnitOption = selectedUtilityOption.energyUnitOptions.find(
           _unitOption => _unitOption.value == utilityEnergyUse.energyUnit);
-        // calculate use
-        if (selectedUtilityOption.isStandardEnergyUnit
-          && selectedUnitOption?.isStandard !== false) {
-          convertedUse = convertValue.convertValue(
-            utilityEnergyUse.energyUse,
-            utilityEnergyUse.energyUnit,
-            companyEnergyUnit).convertedValue;
-        } else {
-          convertedUse = convertValue.convertValue(
-            utilityEnergyUse.energyUse * utilityEnergyUse.energyHHV,
-            utilityEnergyUse.energyUnitStandard,
-            companyEnergyUnit).convertedValue;
-        }
-        if (isNaN(convertedUse)) {
-          convertedUse = 0;
-        }
-        energyUse += convertedUse;
         // calculate cost
         convertedUseForCost = convertValue.convertValue(
           utilityEnergyUse.energyUse,
@@ -113,15 +154,6 @@ export function updateAssessmentUtilityUseCostSavings(assessment: IdbAssessment,
           convertedUseForCost = 0;
         }
         energyCost += convertedUseForCost * facilityUnitSettings[`${camelCaseType}Price`];
-        // calculate saving
-        convertedSaving = convertValue.convertValue(
-          utilityEnergyUse.utilitySaving,
-          utilityEnergyUse.energyUnit,
-          companyEnergyUnit).convertedValue;
-        if (isNaN(convertedSaving)) {
-          convertedSaving = 0;
-        }
-        energySavings += convertedSaving;
         // calculate saving for cost
         convertedSavingForCost = convertValue.convertValue(
           utilityEnergyUse.utilitySaving,
@@ -134,12 +166,9 @@ export function updateAssessmentUtilityUseCostSavings(assessment: IdbAssessment,
       }
     }
   });
-  assessment.energyUse = energyUse;
   assessment.energyCost = energyCost;
   assessment.waterCost = waterCost;
   assessment.cost = energyCost + waterCost;
-  assessment.energySavings = energySavings;
-  assessment.waterSavings = waterSavings;
   assessment.energyCostSavings = energyCostSavings;
   assessment.waterCostSavings = waterCostSavings;
   assessment.costSavings = energyCostSavings + waterCostSavings;
