+++ src/app/shared/shared-assessment-forms/assessment-details-form/assessment-details-form.component.ts
@@ -22,6 +22,7 @@ import { Router } from '@angular/router';
 import { OnSiteVisitIdbService } from 'src/app/indexed-db/on-site-visit-idb.service';
 import { IdbOnSiteVisit } from 'src/app/models/onSiteVisit';
 import { LocaleService } from '../../shared-services/locale.service';
+import { updateAssessmentUtilityUseCostSavings } from '../../reports/calculations/utilityCalculation';
 
 @Component({
     selector: 'app-assessment-details-form',
@@ -137,109 +138,7 @@ export class AssessmentDetailsFormComponent {
 
   async calculateUtilityUseCostSavings() {
     this.updateEnergyOpportunities();
-    let energyUse = 0, energyCost = 0, waterCost = 0;
-    let energySavings = 0, waterSavings = 0;
-    let energyCostSavings = 0, waterCostSavings = 0;
-    this.assessment.utilityTypes.forEach(utilityType => {
-      let utilityEnergyUse: UtilityEnergyUse = this.assessment.utilityEnergyUses.find(
-        _energyUse => _energyUse.utilityType == utilityType);
-      if (utilityEnergyUse.include) {
-        let trimmedType = utilityType.replace(/\s+/g, ''); // Remove spaces
-        let camelCaseType = trimmedType.charAt(0).toLowerCase() + trimmedType.slice(1);
-        let convertedUse = 0, convertedUseForCost = 0;
-        let convertedSaving = 0, convertedSavingForCost = 0;
-        if (utilityType == 'Water' || utilityType == 'Waste Water') {
-          // determine tracked water unit (always use water if water is tracked)
-          this.trackedWaterUnit = utilityEnergyUse.energyUnit;
-          if (utilityType == 'Waste Water') {
-            let waterEnergyUse = this.assessment.utilityEnergyUses.find(
-              _energyUse => _energyUse.utilityType == 'Water');
-            if (waterEnergyUse.include) {
-              this.trackedWaterUnit = waterEnergyUse.energyUnit;
-            }
-          }
-          // calculate cost
-          convertedUseForCost = this.convertValue.convertValue(
-            utilityEnergyUse.energyUse,
-            utilityEnergyUse.energyUnit,
-            this.facilityUnitSettings[`${camelCaseType}Unit`]).convertedValue;
-          if (isNaN(convertedUseForCost)) {
-            convertedUseForCost = 0;
-          }
-          waterCost += convertedUseForCost * this.facilityUnitSettings[`${camelCaseType}Price`];
-          // calculate saving
-          convertedSavingForCost = this.convertValue.convertValue(
-            utilityEnergyUse.utilitySaving,
-            utilityEnergyUse.energyUnit,
-            this.facilityUnitSettings[`${camelCaseType}Unit`]).convertedValue;
-          if (isNaN(convertedSavingForCost)) {
-            convertedSavingForCost = 0;
-          }
-          waterSavings += convertedSavingForCost;
-          waterCostSavings += convertedSavingForCost * this.facilityUnitSettings[`${camelCaseType}Price`];
-        } else {
-          let selectedUtilityOption = UtilityOptions.find(
-            _option => _option.utilityType == utilityType);
-          let selectedUnitOption = selectedUtilityOption.energyUnitOptions.find(
-            _unitOption => _unitOption.value == utilityEnergyUse.energyUnit);
-          // calculate use
-          if (selectedUtilityOption.isStandardEnergyUnit
-            && selectedUnitOption?.isStandard !== false) {
-            convertedUse = this.convertValue.convertValue(
-              utilityEnergyUse.energyUse,
-              utilityEnergyUse.energyUnit,
-              this.companyEnergyUnit).convertedValue;
-            this.trackedEnergyUnit = utilityEnergyUse.energyUnit;
-          } else {
-            convertedUse = this.convertValue.convertValue(
-              utilityEnergyUse.energyUse * utilityEnergyUse.energyHHV,
-              utilityEnergyUse.energyUnitStandard,
-              this.companyEnergyUnit).convertedValue;
-            this.trackedEnergyUnit = utilityEnergyUse.energyUnitStandard;
-          }
-          if (isNaN(convertedUse)) {
-            convertedUse = 0;
-          }
-          energyUse += convertedUse;
-          // calculate cost
-          convertedUseForCost = this.convertValue.convertValue(
-            utilityEnergyUse.energyUse,
-            utilityEnergyUse.energyUnit,
-            this.facilityUnitSettings[`${camelCaseType}Unit`]).convertedValue;
-          if (isNaN(convertedUseForCost)) {
-            convertedUseForCost = 0;
-          }
-          energyCost += convertedUseForCost * this.facilityUnitSettings[`${camelCaseType}Price`];
-          // calculate saving
-          convertedSaving = this.convertValue.convertValue(
-            utilityEnergyUse.utilitySaving,
-            utilityEnergyUse.energyUnit,
-            this.companyEnergyUnit).convertedValue;
-          if (isNaN(convertedSaving)) {
-            convertedSaving = 0;
-          }
-          energySavings += convertedSaving;
-          // calculate saving for cost
-          convertedSavingForCost = this.convertValue.convertValue(
-            utilityEnergyUse.utilitySaving,
-            utilityEnergyUse.energyUnit,
-            this.facilityUnitSettings[`${camelCaseType}Unit`]).convertedValue;
-          if (isNaN(convertedSavingForCost)) {
-            convertedSavingForCost = 0;
-          }
-          energyCostSavings += convertedSavingForCost * this.facilityUnitSettings[`${camelCaseType}Price`];
-        }
-      }
-    });
-    this.assessment.energyUse = energyUse;
-    this.assessment.energyCost = energyCost;
-    this.assessment.waterCost = waterCost;
-    this.assessment.cost = energyCost + waterCost;
-    this.assessment.energySavings = energySavings;
-    this.assessment.waterSavings = waterSavings;
-    this.assessment.energyCostSavings = energyCostSavings;
-    this.assessment.waterCostSavings = waterCostSavings;
-    this.assessment.costSavings = energyCostSavings + waterCostSavings;
+    this.assessment = updateAssessmentUtilityUseCostSavings(this.assessment, this.facilityUnitSettings, this.companyEnergyUnit);
     await this.saveChanges();
   }
